backend/requirements.txt
fastapi
uvicorn
pydantic>=2.0
python-docx
pymupdf
spacy
numpy
sentence-transformers
pandas
